no major changes

+++ NEWTRY/helper/classes.jl
@@ -34,10 +34,10 @@ type OldTrajectory                  # informations about previous trajectories
     cost2target::Array{Float64}     # cost to arrive at the target, i.e. how many iterations from the start to the end of the lap
     curvature::Array{Float64}       # all the curvature calculated in each step of each lap
     oldAlpha::Array{Float64}        # all the alphas computed in each iteration of each LMPC lap
-    costLap::Array{Float64}
+    costLap::Array{Int64}
    
     OldTrajectory(n_oldTraj = 0, oldTraj=Float64[],oldTrajXY=Float64[],oldNIter=Float64[],oldInput=Float64[],costs=Float64[],z_pred_sol=Float64[],
-                  u_pred_sol=Float64[],cost2target= Float64[],curvature=Float64[],oldAlpha=Float64[],costLap=Float64[]) =
+                  u_pred_sol=Float64[],cost2target= Float64[],curvature=Float64[],oldAlpha=Float64[],costLap=Int64[]) =
                  new(n_oldTraj, oldTraj,oldTrajXY,oldNIter,oldInput,costs,z_pred_sol,u_pred_sol,cost2target,curvature,oldAlpha,costLap)
 end
